docs/source/content/developer_documentation/basyx_python_sdk/resources/requirements.txt
basyx-python-sdk
flask
waitress
dash
requests
pytest
isort
pylint
black
